YouLiedEncrypted1
"YouLiedEncrypted1 is an NFT encrypted with bitcoin, giving it its own independent value.

The story is initially encrypted with 0.00041 BTC ($20 at the time of creation). 

Each time an Edition is sold (not resold) $5 BTC will be added to the value of the NFT.

The private key to the story address "14PEk16ngM6dpqvyGtDRhEN2CcaW7iR1jP" is the password to unlock this file. 

The memonic words to the private key are hidden in the story.

Clues to the words may show in other "You Lied" stories.

The 1st to unlock the story gets the bitcoins. Good luck!